JPMS - Fixes Javadoc aggregator-jar build #1022
Change the package of micro-core module to avoid split packages
with the micro module and add module-info to both micro and
arquillian-server modules.
