+++ extension/micro-core/src/main/java/cloud/piranha/micro/core/PiranhaBeanArchiveHandler.java
@@ -25,7 +25,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.micro;
+package cloud.piranha.micro.core;
 
 import static org.jboss.weld.environment.deployment.discovery.jandex.Jandex.INDEX_ATTRIBUTE_NAME;
