+++ extension/micro-core/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-cloud.piranha.micro.CdiExtension
+cloud.piranha.micro.core.CdiExtension
