+++ extension/micro-core/src/main/resources/META-INF/services/org.jboss.weld.environment.deployment.discovery.BeanArchiveHandler
@@ -1,1 +1,1 @@
-cloud.piranha.micro.PiranhaBeanArchiveHandler
+cloud.piranha.micro.core.PiranhaBeanArchiveHandler
