--- micro/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-cloud.piranha.micro.CdiExtension
