--- micro/src/main/resources/META-INF/services/org.jboss.weld.environment.deployment.discovery.BeanArchiveHandler
@@ -1,1 +1,0 @@
-cloud.piranha.micro.PiranhaBeanArchiveHandler
