+++ test/pom.xml
@@ -61,8 +61,7 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
                 <configuration>
-                    <detectOfflineLinks>false</detectOfflineLinks>
-                    <detectLinks>false</detectLinks>
+                    <skip>true</skip>
                 </configuration>
             </plugin>
         </plugins>
